(fix): Correctly Update Nulls Capactity for Nullable Content Stream Data Without Nulls (#331)
Summary:

There are separate chunkers for `NullableContentStreamData` and  `ContentStreamData` streams. However, when a `NullableContentStreamData` stream does not have nulls, [it is treated](https://fburl.com/code/zrsw0dyl) as a `ContentStreamData` stream.  This serves as an optimization when returning `StreamDataView`s since non-nulls will not be materialized or copied. However, there are failures in Vader caused by this code path because in `ContentStreamData` we fail to reserve the needed capacity for null entries.  This diff fixes that.

Reviewed By: helfman

Differential Revision: D87683568