Fix GetRawDatasize Test Util for Nullable Encoding (#339)
Summary:

The test util used to calculate encoding sizes in NimbleDump and VeloxWriter tests only includes the size of the nulls bits for nullable encoded data. This is different from how sizes are nullable content stream sizes are calculated in the rest of our code base and introduced flakiness to a VeloxWriter test. This diff fixes that.

Differential Revision: D88113894