Stream Deduplication (#352)
Summary:
Pull Request resolved: https://github.com/facebookincubator/nimble/pull/352

Introducing stream level deduplication.

With this change, before writing streams to a stripe, we are identifying streams that are byte-identical and only write one copy of the stream to the file.
All other identical streams will store the same offset and size as the original stream.

To opt-in for this behavior, a new option was introduced: streamDeduplicationEnabled.

C++ reader was modified to identify if streams are duplicated and only load them once.
Selective reader seems to already handle this read deduplication.
Java was not modified. It is loading the same stream multiple times (which is not a regression over what we currently have).

Reviewed By: sdruzkin

Differential Revision: D88087515

fbshipit-source-id: 1ed10ea0b1819cf7eb5a18fc08609024640c2367