Fork encodings for new API release (#361)
Summary:
Pull Request resolved: https://github.com/facebookincubator/nimble/pull/361

We need to make a breaking change to the `Encoding` interface to allow us to pass string buffers directly from Encodings to the decoders and column/field readers. In order for a safe release without introducing excessive branching in the current implementation, we create a fallback version of the current encodings and create knobs to invoke the different versions of factories to create them.

Reviewed By: helfman

Differential Revision: D87965117

fbshipit-source-id: 101522ee3c6e27e637496f125e32c133767ddaac