misc: Respect Linux cgroups restrictions when sizing thread pools (#367)
Summary:
Pull Request resolved: https://github.com/facebookincubator/nimble/pull/367

This change replaces uses of std::thread::hardware_concurrency with
folly::hardware_concurrency the latter of which always respects
container restrictions while the former does not.  This ensures that
when running inside of a Linux container with restrictions on the
available processors, thread pools and other data structures will not
be over-provisioned.

Reviewed By: sdruzkin

Differential Revision: D87913690

fbshipit-source-id: 345ca99c996867274a2f4d5196e1ea513a6850a9